% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AsbestosData}
\alias{AsbestosData}
\title{Asbestos data}
\format{A contingency table with 5 rows and 4 columns
\describe{
  \item{rows}{Years working with asbestos: 0-9, 10-19, 20-29, 30-39, 40+}
  \item{columns}{Has asbestosis or not: No (N), severity graded as 1-3 (G1-G3)}
}}
\source{
Still looking for it
}
\usage{
AsbestosData
}
\description{
Cases and severity of asbestosis, classified by years working with asbestos
}
\keyword{datasets}
