% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AttachmentData}
\alias{AttachmentData}
\title{van Ijzendoorn's attachment data}
\format{
A contingency table with 4 rows and 4 columns
\describe{
  \item{rows}{Infant response: Avoidant, Secure, Resistant, Disorganised}
  \item{columns}{Mother's Classification: Dismissing, Autonomous, Preoccupied, Unresolved}
}
}
\source{
E.J. Beh,
\emph{Elliptical confidence regions for simple correspondence analysis},
Journal of Statistical Planning and Inference 140 (2010), pp. 2582--2588.
}
\usage{
AttachmentData
}
\description{
Classification of mother's attachment to her child and child's reaction
}
\keyword{datasets}
