
\name{DreamData}
     \docType{data}
     \alias{DreamData}
     \title{Maxwell's Dream Data}
     \description{
       A cross-classification of counts of boys by age-group (A=5-7, B=8-9, C=10-11, D=12-13, E=14-15)
       and severity of dream-disturbance (a=lowest, d=highest).
     }
     \usage{DreamData}
     \format{A contingency table.}
     \source{Iliopoulos et al.}
     \references{
       G. Iliopoulos, M. Kateri and I. Ntzoufras (2007). 
       \emph{Bayesian estimation of unrestricted and order-restricted association models for a two-way contingency table}.
       Comput. Statist. Data Anal., 51, 4643-4655. 
     }
     \keyword{datasets}