% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{getindicator}
\alias{getindicator}
\title{Converting a data matrix into an indicator matrix}
\usage{
getindicator(
  Xinput,
  input = "nbyp",
  Jk = NULL,
  maxcat = NULL,
  varandcat = TRUE
)
}
\arguments{
\item{Xinput}{A data matrix, in the form of a data frame or similar}

\item{input}{See \code{\link{convert}}}

\item{Jk}{See \code{\link{convert}}}

\item{maxcat}{See \code{\link{convert}}}

\item{varandcat}{See \code{\link{convert}}}
}
\value{
An indicator matrix, where a variable with J_k categories is represented by J_k columns and a
data point taking the i-th category has 1 in the i-th of these columns and a zero in the others
}
\description{
\code{getindicator} recodes a data matrix from one format (n objects by p variables,
counts for distinct combinations of p variables, contingency table) into an indicator matrix
}
\examples{
nishindicator <- getindicator(NishData)

}
\seealso{
Other conversion functions: 
\code{\link{convert}()},
\code{\link{getBurt}()},
\code{\link{getCT}()},
\code{\link{getdoubled}()}
}
\concept{conversion functions}
