% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{OsteoDataNames}
\alias{OsteoDataNames}
\title{Osteoarchaeological data with named categories}
\format{
A matrix of 6027 bones by 11 variables, each with 2-12 categories
\describe{
  \item{Species/Taxon - Taxa}{Sheep/goat,Pig,Cattle,Deer,Dog,Equid,Large,Medium,Small,Size:indet}
  \item{Anatomical element - Elem}{Horn/antler,Head,Neck,Axial,Upper_front,Lower_front,Upper_hind,Lower_hind,Feet,Metapodials,Long_bone,Element:indet}
  \item{Recent breakage - Rec}{P/A}
  \item{Cut marks - Cut}{P/A}
  \item{Gnawing - Gnaw}{P/A}
  \item{Weathering stage - Weat}{stage 1-4 or A}
  \item{Thermal stage - Therm}{stage 1-5 or A}
  \item{Trampling - Tram}{P/A}
  \item{Root etching - Root}{P/A}
  \item{Post-depositional - Post}{P/A}
  \item{Context - Context}{Secondary,Room_fills,Primary,Floors}
}
}
\source{
Macheridis, S.
\emph{The Use of Multiple Correspondence Analysis (MCA) in Taphonomy: The Case of Middle Helladic Asine, Greece}
International Journal of Osteoarchaeology 27 (2017), pp. 477–-487.
}
\usage{
OsteoDataNames
}
\description{
Animal bones classified by 11 different variables
}
\keyword{datasets}
