% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineclassesandfunctions.R
\name{printca}
\alias{printca}
\title{Prints reasonably full results, including variances}
\usage{
printca(x, datasetname = NULL)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{cabootcrsresults}}}

\item{datasetname}{The name (in "") of the data set, to be used in the output, defaults to name in cabootcrs object}
}
\value{
Printed results, no plots or objects produced
}
\description{
\code{printca} prints full correspondence analysis results, including inertias, coordinates,
representations, contributions, variances, covariances and critical values
}
\examples{
results <- cabootcrs(DreamData, showresults=FALSE)
printca(results, datasetname="Dreams")

}
\seealso{
\code{\link{cabootcrs-package}}, \code{\link{cabootcrs}}, \code{\link{summaryca}}, \code{\link{plotca}},
\code{\linkS4class{cabootcrsresults}}
}
