% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineclassesandfunctions.R
\name{settingsinertias}
\alias{settingsinertias}
\title{Internal function to be used by printca and summaryca}
\usage{
settingsinertias(x)
}
\arguments{
\item{x}{An object of class \code{\linkS4class{cabootcrsresults}}}
}
\value{
printed output only
}
\description{
\code{settingsinertias} prints the settings and the inertias
}
\examples{
# Purely internal, not intended for use by users

}
\seealso{
\code{\link{summaryca}}, \code{\link{printca}}
}
