% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineclassesandfunctions.R
\docType{class}
\name{cabasicresults-class}
\alias{cabasicresults-class}
\title{A class containing the basic results from CA}
\description{
This is intended for internal use within \code{\link{cabootcrs}} and only contains the
data structures required for each bootstrap replicate
}
\section{Slots}{

\describe{
\item{\code{Rprofile}}{Row profile matrix, class \code{"matrix"}}

\item{\code{Cprofile}}{Column profile matrix, class \code{"matrix"}}

\item{\code{Rweights}}{Matrix of weights for row points: square roots of inverse column sums, class \code{"matrix"}}

\item{\code{Cweights}}{Matrix of weights for column points: square roots of inverse row sums, class \code{"matrix"}}

\item{\code{Raxes}}{Matrix of axes for row points: right singular vectors of weighted, centred data matrix, class \code{"matrix"}}

\item{\code{Caxes}}{Matrix of axes for column points: left singular vectors of weighted, centred data matrix, class \code{"matrix"}}

\item{\code{r}}{Rank of weighted, centred data matrix, class \code{"numeric"}}

\item{\code{realr}}{In multiple CA, the number of singular values (Burt matrix) or squared singular values
(indicator matrix) exceeding \eqn{1/p} where \eqn{p} is the number of variables,
class \code{"numeric"}}

\item{\code{mu}}{Singular values of weighted, centred data matrix, class \code{"numeric"}}
}}

\seealso{
\code{\linkS4class{cabootcrsresults}}
}
