% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DreamData}
\alias{DreamData}
\title{Maxwell's dream data set}
\format{A contingency table with 5 rows and 4 columns
\describe{
  \item{rows}{Age group: 5-7 (A), 8-9 (B), 10-11 (C), 12-13 (D), 13-14 (E)}
  \item{columns}{Severity of disturbance of dream: lowest (a) to highest (d)} 
}}
\source{
G. Iliopoulos, M. Kateri and I. Ntzoufras,
\emph{Bayesian estimation of unrestricted and order-restricted association
models for a two-way contingency table},
Computational Statistics and Data Analysis 51 (2007), pp. 4643--4655.
}
\usage{
DreamData
}
\description{
Reported disturbance of dreams among boys
}
\keyword{datasets}
